backend/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
Authlib==1.6.5
bandit==1.8.6
bcrypt==5.0.0
black==25.9.0
boto3==1.40.57
botocore==1.40.57
build==1.3.0
CacheControl==0.14.3
celery==5.4.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.0
colorama==0.4.6
coverage==7.11.0
crashtest==0.4.1
cryptography==46.0.3
distlib==0.4.0
dnspython==2.8.0
dparse==0.6.4
dulwich==0.24.6
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.0
flake8==7.3.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.11
iniconfig==2.3.0
installer==0.7.0
isort==7.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
keyring==25.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.9
more-itertools==10.8.0
msgpack==1.1.2
mypy==1.18.2
mypy_extensions==1.1.0
nltk==3.9.2
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pbs-installer==2025.10.14
pkginfo==1.12.1.2
platformdirs==4.5.0
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
pre_commit==4.3.0
psutil==7.1.1
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.2.0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.1
redis==7.0.0
Pillow==10.4.0
ffmpeg-python==0.2.0
prometheus-client==0.20.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
safety==3.6.2
safety-schemas==0.0.16
sentry-sdk==2.42.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
stevedore==5.5.0
structlog==25.4.0
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
trove-classifiers==2025.9.11.17
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.3
watchfiles==1.1.1
websockets==15.0.1
zstandard==0.25.0
